# calendly-clone-api/docs/en/api.yaml
openapi: 3.0.0
info:
  title: Calendly Clone API
  description: API for scheduling and managing appointments (English documentation)
  version: 1.0.0
servers:
  - url: /
    description: API endpoints
paths:
...
  /events:
    post:
      security:
        - BearerAuth: []
      tags:
        - Events
      summary: Create a new event type
      description: Create a new event type with a name, duration, and optional details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventType'
            example:
              name: "30 Minute Meeting"
              duration: 30
              description: "Short consultation call"
              color: "#4CAF50"
      responses:
        '201':
          description: Event type created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
              example:
                id: "evt-123e4567-e89b-12d3-a456-426614174000"
                name: "30 Minute Meeting"
                duration: 30
                description: "Short consultation call"
                color: "#4CAF50"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid input"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      security:
        - BearerAuth: []
      tags:
        - Events
      summary: List all events
      description: Retrieve a list of all event types.
      responses:
        '200':
          description: List of events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
              example:
                - id: "evt-123e4567-e89b-12d3-a456-426614174000"
                  name: "30 Minute Meeting"
                  duration: 30
                  description: "Short consultation call"
                  color: "#4CAF50"
                - id: "evt-223e4567-e89b-12d3-a456-426614174001"
                  name: "60 Minute Meeting"
                  duration: 60
                  description: "In-depth discussion"
                  color: "#2196F3"
                - id: "evt-323e4567-e89b-12d3-a456-426614174002"
                  name: "15 Minute Coffee Chat"
                  duration: 15
                  description: "Quick catch-up"
                  color: "#FF9800"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{eventId}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Events
      summary: Get event details
      description: Retrieve details of a specific event by its ID.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the event to retrieve.
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              example:
                id: "evt-123e4567-e89b-12d3-a456-426614174000"
                name: "30 Minute Meeting"
                duration: 30
                description: "Short consultation call"
                color: "#4CAF50"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required"
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Event with ID evt-123e4567 not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      security:
        - BearerAuth: []
      tags:
        - Events
      summary: Partially update an event
      description: Update specific fields of an event by its ID.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the event to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPatch'
            example:
              name: "45 Minute Meeting"
              duration: 45
              description: "Extended consultation call"
              color: "#9C27B0"
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
              example:
                id: "evt-123e4567-e89b-12d3-a456-426614174000"
                name: "45 Minute Meeting"
                duration: 45
                description: "Extended consultation call"
                color: "#9C27B0"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Duration must be a positive number"
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: []
      tags:
        - Events
      summary: Delete an event
      description: Delete a specific event by its ID.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the event to delete.
      responses:
        '204':
          description: Event deleted successfully
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Resource not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    EventType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        duration:
          type: integer
        description:
          type: string
        color:
          type: string
      required:
        - name
        - duration
...
    Event:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        duration:
          type: integer
        description:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - duration
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Bad Request"
      required:
        - message
...
    EventPatch:
      type: object
      properties:
        name:
          type: string
        duration:
          type: integer
        description:
          type: string
        color:
          type: string
      minProperties: 1  # At least one field must be provided
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token